application/requirements.txt
streamlit>=1.28
langchain>=0.0.217
openai>=1.2
duckduckgo-search
anthropic>=0.3.0
trubrics>=1.4.3
streamlit-feedback
streamlit-authenticator
st-pages
python-dotenv~=1.0.1
openpyxl~=3.1.0

PyYAML~=6.0.1
pandas~=2.2.2
PyPDF2